TP4/HIT2Nuevo/workers/requirements.txt
pika
opencv-python
numpy
redis
requests
google-cloud-storage